scranton/agents/requirements.txt
google-adk>=1.3.0
google-genai>=1.20.0
python-dotenv>=1.1.0
litellm>=1.72.7
sqlalchemy>=2.0.41
psycopg2-binary>=2.9.10
pydantic>=2.11.7